refactor: a refactor
